Extract BreadCrumb component
