+++ src/modules/core/components/ExpandedParagraph/ExpandedParagraph.tsx
@@ -17,15 +17,36 @@ const MSG = defineMessages({
 });
 
 interface Props {
+  /*
+   * This contains a long text paragraph that initially gets shown
+   * in a shortened version that can be extended clicking a more button
+   */
   paragraph: string;
+  /*
+   * In case the developer wants to add further nodes to the shortened
+   * version of the paragraph, this is the place to do it.
+   */
   elements?: ReactNode;
+  /*
+   * This is to add react nodes to the extended version of the
+   * paragraph.
+   */
   expandedElements?: ReactNode;
+  /*
+   * This is the number of characters that the shortened version of the
+   * paragraph will have.
+   */
   characterLimit: number;
+  /*
+   * This is the number of characters that the extended version of the
+   * paragraph will have.
+   */
   maximumCharacters: number;
 }
 
 const ExpandedParagraph = ({
   paragraph,
+  elements,
   expandedElements,
   characterLimit,
   maximumCharacters,
@@ -37,6 +58,7 @@ const ExpandedParagraph = ({
         paragraph,
         expanded ? maximumCharacters : characterLimit,
       )}
+      {elements && elements}
       {!expanded && (
         <Button
           onClick={() => {
